Add buffer local variable for balancing all brackets
paredit balanced brackets are addicting once you get used to it.

When editing JavaScript, Terraform, or various other languages,
there are times when one finds oneself trying to slurp, barf,
and have balanced brackets.

This adds a new buffer local flag, and initializer which adding
the ability to have balancing brackets in languages besides clojure,
scheme, racket, lfe, etc. by letting the user init the buffer
for a filetype by calling a new initializer
PareditInitBalancingAllBracketsBuffer to enable bracket balancing
as they see fit, such as with an autocommand on a given filetype.